fix output ESP32 & ESP 8266
- fix, no calling of commands possible
