--- src/arduino-ide/SIGNALDuino/output.cpp
@@ -1,1 +1,0 @@
-../../_micro-api/libraries/output/src/output.cpp
